dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_event.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

'use strict';

SimpleTest.waitForExplicitFinish();

function runTest() {
  ok(navigator.fileSystemProvider, 'should have navigator.fileSystemProvider');

  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv({"set": [["device.storage.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{'type': 'filesystemprovider', 'allow': true, 'context': document}], runTest);
});

 </script>
 </pre>
 </body>
 </html>
